Fix dynamic condition proccing too many times on passive damage buffs, add ability to proc multiple ConsumeBuffs from one trigger
If a dynamic condition is added to a Passive Damage buff (e.g. forte generation on Yuanwu's Coordinated Attack), there was a bug where the dynamic condition would be evaluated once when creating the buff and once on the initial proc, which happens at the same time. Also, evaluating the dynamic condition on proc now takes into account the number of times the Passive Damage was procced.

Also changed removeBuff to a list, so that if one skill triggers multiple different ConsumeBuff buffs, they can all be removed.